requirements.txt
aiohttp
botbuilder-core
botbuilder-schema
